Get dockerized backend serving on linux
In my attempts to get the dockerized backend up and running on my linux
machine, I encountered this issue where nginx did not have access to the
shared volume files. It turns out that volumes exposed from a linux host
machine's file system abide by the host system's ACL.

I've resolved this issue by first copying the web server files into the
app container, changing the permissions as needed, and then creating a
volume that is only shared between the nginx and app containers. This
way, the containers control the file system ACL and the host filesystem
is out of the way.

This potentially complicates processes involving bumping dependencies
but otherwise seems like a pretty clean solution, since we don't need to
modify the nginx image at all.

Definitely let me know if I've broken a dependency
management workflow, or any other workflow for that matter, in which
case the nginx image user can be modified to work nicely with the host.

test plan:

1. follow the setup instructions in the markdown file
2. request http://localhost:900 once the containers are all up
3. Observe the behavior matches that of the staging endpoint when
   queried: https://staging-api.coordinape.com/api