Get dockerized backend serving on linux: DevX Friendly version
In my attempts to get the dockerized backend up and running on my linux
machine, I encountered this issue where nginx did not have access to the
shared volume files. It turns out that volumes exposed from a linux host
machine's file system force containers to abide by the host system's ACL.

In an effort to preserve hot-reload functionality while running this
suite of containers runs, I've left the host-exposed docker volume setup
as-is and instead updated the nginx container to have the same UID and
GID for the nginx user as the host file system. These IDs can now be
modified inside the docker-compose file if needed as well.

test plan:

1. follow the setup instructions in the markdown file
2. request http://localhost:9000 once the containers are all up
3. Observe the behavior matches that of the staging endpoint when
   queried: https://staging-api.coordinape.com/api